+++ server/mock/store.go
@@ -6,6 +6,7 @@ import (
 	dbm "github.com/cosmos/cosmos-db"
 	protoio "github.com/cosmos/gogoproto/io"
 
+	"github.com/cosmos/cosmos-sdk/store/metrics"
 	pruningtypes "github.com/cosmos/cosmos-sdk/store/pruning/types"
 	snapshottypes "github.com/cosmos/cosmos-sdk/store/snapshots/types"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
@@ -50,6 +51,10 @@ func (ms multiStore) SetTracer(w io.Writer) storetypes.MultiStore {
 }
 
 func (ms multiStore) AddListeners(key storetypes.StoreKey, listeners []storetypes.WriteListener) {
+	panic("not implemented")
+}
+
+func (ms multiStore) SetMetrics(metrics.StoreMetrics) {
 	panic("not implemented")
 }