+++ store/store.go
@@ -7,13 +7,14 @@ import (
 	"github.com/tendermint/tendermint/libs/log"
 
 	"github.com/cosmos/cosmos-sdk/store/cache"
+	"github.com/cosmos/cosmos-sdk/store/metrics"
 	"github.com/cosmos/cosmos-sdk/store/rootmulti"
 	"github.com/cosmos/cosmos-sdk/store/types"
 	"github.com/cosmos/cosmos-sdk/types/kv"
 )
 
-func NewCommitMultiStore(db dbm.DB, logger log.Logger) types.CommitMultiStore {
-	return rootmulti.NewStore(db, logger)
+func NewCommitMultiStore(db dbm.DB, logger log.Logger, metricGatherer metrics.StoreMetrics) types.CommitMultiStore {
+	return rootmulti.NewStore(db, logger, metricGatherer)
 }
 
 func NewCommitKVStoreCacheManager() types.MultiStorePersistentCache {